Aesthetics and fixed a bug with singleWordTranslation
Made tag <p> !important
and changed the color on background for rgba

Does !important directly change everything on the page? Or does it keep
our code from being messed up from each website?

Just wanted to fix up singleWordTranslation to be more consistent, but
we can remove it if it becomes too distracting.